+++ src/decompress/zstd_decompress_block.rs
@@ -4334,11 +4334,11 @@ unsafe extern "C" fn ZSTD_decodeSequence(
         }
         (*seqState)
             .prevOffset[2 as std::ffi::c_int
-            as usize] = (*seqState).prevOffset[1 as usize];
+            as usize] = (*seqState).prevOffset[1];
         (*seqState)
             .prevOffset[1 as std::ffi::c_int
-            as usize] = (*seqState).prevOffset[0 as usize];
-        (*seqState).prevOffset[0 as usize] = offset;
+            as usize] = (*seqState).prevOffset[0];
+        (*seqState).prevOffset[0] = offset;
     } else {
         let ll0 = ((*llDInfo).baseValue == 0)
             as std::ffi::c_int as u32;
@@ -4350,7 +4350,7 @@ unsafe extern "C" fn ZSTD_decodeSequence(
                 .prevOffset[1 as std::ffi::c_int
                 as usize] = (*seqState)
                 .prevOffset[(ll0 == 0) as std::ffi::c_int as usize];
-            (*seqState).prevOffset[0 as usize] = offset;
+            (*seqState).prevOffset[0] = offset;
         } else {
             offset = (ofBase.wrapping_add(ll0) as usize)
                 .wrapping_add(
@@ -4360,7 +4360,7 @@ unsafe extern "C" fn ZSTD_decodeSequence(
                     ),
                 );
             let mut temp = if offset == 3 {
-                ((*seqState).prevOffset[0 as usize])
+                ((*seqState).prevOffset[0])
                     .wrapping_sub(1)
             } else {
                 (*seqState).prevOffset[offset as usize]
@@ -4369,13 +4369,13 @@ unsafe extern "C" fn ZSTD_decodeSequence(
             if offset != 1 {
                 (*seqState)
                     .prevOffset[2 as std::ffi::c_int
-                    as usize] = (*seqState).prevOffset[1 as usize];
+                    as usize] = (*seqState).prevOffset[1];
             }
             (*seqState)
                 .prevOffset[1 as std::ffi::c_int
-                as usize] = (*seqState).prevOffset[0 as usize];
+                as usize] = (*seqState).prevOffset[0];
             offset = temp;
-            (*seqState).prevOffset[0 as usize] = offset;
+            (*seqState).prevOffset[0] = offset;
         }
     }
     seq.offset = offset;
